refactor: remove unused auth routes
