+++ src/auth/authController.js
@@ -1,28 +1,21 @@
-const { omit } = require('lodash');
 const { RESPONSE_MESSAGE } = require('../../constant');
 const { ConflictError, UnAuthorizedError } = require('../../utils/customError');
-const { createToken, validateCredentials } = require('../../utils/helper');
-const accountsService = require('../accounts/accountsService');
 const responseHandler = require('../../utils/responseHandler');
+const authService = require('./authService');
 
 const registerAccount = async (req, res) => {
   const payload = { ...req.body };
-  let account = await accountsService.accountExists(payload.email);
+  let registrationData = await authService.registerAccount(payload);
 
-  if (account) {
+  if (!registrationData) {
     throw new ConflictError(RESPONSE_MESSAGE.CONFLICT(payload.category));
   }
+  const { token, user } = registrationData;
 
-  account = await accountsService.createAccount(payload);
-  account = omit(account.toObject(), 'password');
-
-  const token = createToken({
-    id: account._id,
-  });
-
+  res.cookie('jwt_token', token);
   new responseHandler(
     res,
-    { token, user: account },
+    { token, user },
     201,
     RESPONSE_MESSAGE.CREATE_SUCCESSFUL(payload.category)
   );
@@ -31,15 +24,13 @@ const registerAccount = async (req, res) => {
 const userLogin = async (req, res) => {
   const { email, password } = req.body;
 
-  let account = await validateCredentials(email, password);
-  if (!account) {
+  let loginData = await authService.userLogin({ email, password });
+  if (!loginData) {
     throw new UnAuthorizedError('Invalid credentials!');
   }
-  const token = createToken({
-    id: account._id,
-  });
+  const { token, user } = loginData;
 
-  const user = omit(account.toObject(), 'password', '__v');
+  res.cookie('jwt_token', token);
   new responseHandler(res, { token, user }, 200, RESPONSE_MESSAGE.SUCCESS);
 };
 
@@ -48,10 +39,7 @@ const userLogout = async (req, res) => {
     req.logout();
   }
   res.clearCookie('jwt_token');
-  res.status(200).json({
-    statusCode: 200,
-    message: RESPONSE_MESSAGE.LOGOUT,
-  });
+  new responseHandler(res, undefined, 200, RESPONSE_MESSAGE.LOGOUT);
 };
 
 module.exports = {
