+++ src/auth/authValidator.js
@@ -15,7 +15,7 @@ const loginValidator = Joi.object({
   password: Joi.string()
     .required()
     .min(8)
-    .pattern(new RegExp('^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%^&*])'))
+    .pattern(REGEX_PATTERNS.PASSWORD)
     .messages({
       'string.pattern.base':
         'password must contain uppercase, lowercase, number and special character',
@@ -24,76 +24,6 @@ const loginValidator = Joi.object({
       'any.required': 'Password is required',
     }),
 }).with('email', 'password');
-
-const createStudentValidator = Joi.object({
-  firstName: Joi.string().required().messages({
-    'string.empty': 'Firstname is required',
-    'any.required': 'Firstname is required',
-  }),
-  lastName: Joi.string().required().messages({
-    'string.empty': 'Lastname is required',
-    'any.required': 'Lastname is required',
-  }),
-  email: Joi.string().email().required().messages({
-    'string.email': 'Not a valid email address',
-    'string.empty': 'Email is required',
-    'any.required': 'Email is required',
-  }),
-  password: Joi.string()
-    .required()
-    .min(8)
-    .pattern(new RegExp('^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%^&*])'))
-    .messages({
-      'string.pattern.base':
-        'password must contain uppercase, lowercase, number and special character',
-      'string.min': 'Password must be at least 8 characters long',
-      'string.empty': 'Password is required',
-      'any.required': 'Password is required',
-    }),
-  track: Joi.string().required().messages({
-    'string.empty': 'Track is required',
-    'any.required': 'Track is required',
-  }),
-  matric: Joi.string().allow('').optional(),
-  altSchoolId: Joi.string().allow('').optional(),
-  category: Joi.string().allow('').optional(),
-  stack: Joi.string().allow('').optional(),
-  gender: Joi.string().allow('').optional(),
-});
-
-const createMentorValidator = Joi.object({
-  firstName: Joi.string().required().messages({
-    'string.empty': 'Firstname is required',
-    'any.required': 'Firstname is required',
-  }),
-  lastName: Joi.string().required().messages({
-    'string.empty': 'Lastname is required',
-    'any.required': 'Lastname is required',
-  }),
-  email: Joi.string().email().required().messages({
-    'string.email': 'Not a valid email address',
-    'string.empty': 'Email is required',
-    'any.required': 'Email is required',
-  }),
-  password: Joi.string()
-    .required()
-    .min(8)
-    .pattern(new RegExp('^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%^&*])'))
-    .messages({
-      'string.pattern.base':
-        'password must contain uppercase, lowercase, number and special character',
-      'string.min': 'Password must be at least 8 characters long',
-      'string.empty': 'Password is required',
-      'any.required': 'Password is required',
-    }),
-  track: Joi.string().required().messages({
-    'string.empty': 'Track is required',
-    'any.required': 'Track is required',
-  }),
-  category: Joi.string().allow('').optional(),
-  specialization: Joi.string().allow('').optional(),
-  yearsOfExperience: Joi.number().allow('').optional(),
-});
 
 const createAccountValidator = Joi.object({
   firstName: Joi.string().required().messages({
@@ -144,7 +74,5 @@ const createAccountValidator = Joi.object({
 
 module.exports = {
   createAccountValidator,
-  createMentorValidator,
-  createStudentValidator,
   loginValidator,
 };
